Add ability to exclude file names from VFS sync
This patch is for mitigating the inability of the OpenCms VFS sync to
exclude VCS metadata. The opencms-workplace.xml now supports two
additional tags:
<workplace>
  [...]
  <synchronization>
    <exclude-pattern>^.</exclude-pattern>
    <exclude-pattern>^CVS$</exclude-pattern>
    [...]
  </synchronization>
</workplace>
The patterns are parsed as Java RegExes and will neither be imported to
the VFS nor deleted from the RFS by the synchronization process.

This Patch is missing unit tests and only contains the english and
german localization.